requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
huggingface-hub>=0.19.4 